+++ auction-server/src/api.rs
@@ -69,6 +69,10 @@ use {
     },
     spl_associated_token_account::instruction::AssociatedTokenAccountInstruction,
     std::{
+        fmt::{
+            Display,
+            Formatter,
+        },
         sync::{
             atomic::Ordering,
             Arc,
@@ -167,21 +171,21 @@ impl std::fmt::Display for InstructionError {
                 write!(
                     f,
                     "Invalid from account in sol transfer instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidToAccountTransferInstruction { expected, found } => {
                 write!(
                     f,
                     "Invalid to account in sol transfer instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidAmountTransferInstruction { expected, found } => {
                 write!(
                     f,
                     "Invalid amount in sol transfer instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidSyncNativeInstructionCount(address) => {
@@ -208,42 +212,42 @@ impl std::fmt::Display for InstructionError {
                 write!(
                     f,
                     "Invalid destination account in close account instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidOwnerCloseAccountInstruction { expected, found } => {
                 write!(
                     f,
                     "Invalid owner in close account instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidMintInCreateAtaInstruction { expected, found } => {
                 write!(
                     f,
                     "Invalid mint in create associated token account instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidOwnerInCreateAtaInstruction { expected, found } => {
                 write!(
                     f,
                     "Invalid owner in create associated token account instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidPayerInCreateAtaInstruction { expected, found } => {
                 write!(
                     f,
                     "Invalid payer in create associated token account instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidTokenProgramInCreateAtaInstruction { expected, found } => {
                 write!(
                     f,
                     "Invalid token program in create associated token account instruction. Expected: {:?} found: {:?}",
-                    found, expected
+                    expected, found
                 )
             }
             InstructionError::InvalidSystemProgramInCreateAtaInstruction(program) => {
@@ -564,6 +568,12 @@ impl IntoResponse for RestError {
     fn into_response(self) -> Response {
         let (status, msg) = self.to_status_and_message();
         (status, Json(ErrorBodyResponse { error: msg })).into_response()
+    }
+}
+
+impl Display for RestError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.to_status_and_message().1)
     }
 }