+++ auction-server/src/auction/service/submit_quote.rs
@@ -36,7 +36,7 @@ impl Service<Svm> {
         let auction: entities::Auction<Svm> = self
             .get_auction_by_id(GetAuctionByIdInput { auction_id })
             .await
-            .ok_or(RestError::BadParameters("Invalid quote".to_string()))?;
+            .ok_or(RestError::BadParameters("Quote not found. The provided reference ID may be invalid, already finalized on-chain, or canceled.".to_string()))?;
 
         let winner_bid = auction
             .bids
@@ -47,11 +47,11 @@ impl Service<Svm> {
                     || bid.status.is_cancelled()
             })
             .cloned()
-            .ok_or(RestError::BadParameters("Invalid quote".to_string()))?;
+            .ok_or(RestError::BadParameters("This quote has already been submitted and finalized on-chain. No further changes are allowed.".to_string()))?;
 
         if winner_bid.status.is_submitted() {
             Err(RestError::BadParameters(
-                "Quote is already submitted".to_string(),
+                "Quote is already submitted on-chain.".to_string(),
             ))
         } else if winner_bid.status.is_cancelled() {
             Err(RestError::BadParameters(
@@ -121,6 +121,7 @@ impl Service<Svm> {
         Ok(bid.chain_data.transaction)
     }
 
+    #[tracing::instrument(skip_all, err)]
     pub async fn submit_quote(
         &self,
         input: SubmitQuoteInput,
@@ -133,23 +134,25 @@ impl Service<Svm> {
                 bid.chain_data.transaction.clone(),
                 entities::BidPaymentInstructionType::Swap,
             )
-            .map_err(|_| RestError::BadParameters("Invalid quote".to_string()))?;
+            .map_err(|_| RestError::BadParameters("Invalid quote.".to_string()))?;
         let SwapAccounts { user_wallet, .. } = self
             .extract_swap_accounts(&bid.chain_data.transaction, &swap_instruction)
             .await
-            .map_err(|_| RestError::BadParameters("Invalid quote".to_string()))?;
+            .map_err(|_| RestError::BadParameters("Invalid quote.".to_string()))?;
         let swap_args = Self::extract_swap_data(&swap_instruction)
-            .map_err(|_| RestError::BadParameters("Invalid quote".to_string()))?;
+            .map_err(|_| RestError::BadParameters("Invalid quote.".to_string()))?;
 
         if swap_args.deadline < (OffsetDateTime::now_utc() - DEADLINE_BUFFER).unix_timestamp() {
-            return Err(RestError::BadParameters("Quote is expired".to_string()));
+            return Err(RestError::BadParameters("Quote is expired.".to_string()));
         }
 
         if !input.user_signature.verify(
             &user_wallet.to_bytes(),
             &bid.chain_data.transaction.message.serialize(),
         ) {
-            return Err(RestError::BadParameters("Invalid signature".to_string()));
+            return Err(RestError::BadParameters(
+                "Invalid user signature.".to_string(),
+            ));
         }
 
         let user_signature_pos = bid
@@ -163,7 +166,7 @@ impl Service<Svm> {
         bid.chain_data.transaction.signatures[user_signature_pos] = input.user_signature;
         if bid.chain_data.bid_payment_instruction_type != entities::BidPaymentInstructionType::Swap
         {
-            return Err(RestError::BadParameters("Invalid quote".to_string()));
+            return Err(RestError::BadParameters("Invalid quote.".to_string()));
         }
 
         self.sign_bid_and_submit_auction(bid, auction, true).await