+++ auction-server/src/opportunity/entities/opportunity_svm.rs
@@ -19,9 +19,9 @@ use {
     },
     ::express_relay::FeeToken as ProgramFeeToken,
     express_relay::state::FEE_SPLIT_PRECISION,
-    express_relay_api_types::{
-        opportunity as api,
-        opportunity::QuoteTokensWithTokenPrograms,
+    express_relay_api_types::opportunity::{
+        self as api,
+        QuoteTokensWithTokenPrograms,
     },
     serde::{
         Deserialize,
@@ -29,8 +29,11 @@ use {
     },
     solana_sdk::{
         clock::Slot,
+        program_pack::Pack,
         pubkey::Pubkey,
+        rent::Rent,
     },
+    spl_token_2022::state::Account as TokenAccount,
     std::ops::Deref,
     time::{
         Duration,
@@ -274,6 +277,28 @@ pub fn get_opportunity_swap_data(opp: &OpportunitySvm) -> &OpportunitySvmProgram
         _ => {
             panic!("Opportunity must be a swap opportunity to get swap data");
         }
+    }
+}
+
+impl OpportunitySvmProgramSwap {
+    pub fn get_user_amount_to_wrap(&self, amount_user: u64) -> u64 {
+        let number_of_atas_paid_by_user = [
+            &self.token_account_initialization_configs.user_ata_mint_user,
+            &self
+                .token_account_initialization_configs
+                .user_ata_mint_searcher,
+        ]
+        .iter()
+        .filter(|&&config| matches!(config, TokenAccountInitializationConfig::UserPayer))
+        .count();
+
+        std::cmp::min(
+            amount_user,
+            self.user_mint_user_balance.saturating_sub(
+                number_of_atas_paid_by_user as u64
+                    * Rent::default().minimum_balance(TokenAccount::LEN), // todo: token2022 accounts can be bigger than this, this hack might not work for them
+            ),
+        )
     }
 }