+++ sdk/js/src/index.ts
@@ -727,6 +727,7 @@ export class Client {
             opportunity.token_account_initialization_configs.user_ata_mint_user,
         },
         memo: opportunity.memo ?? undefined,
+        userMintUserBalance: new anchor.BN(opportunity.user_mint_user_balance),
       };
     } else {
       console.warn("Unsupported opportunity", opportunity);