+++ sdk/rust/src/svm.rs
@@ -28,7 +28,9 @@ use {
         clock::Slot,
         hash::Hash,
         instruction::Instruction,
+        program_pack::Pack,
         pubkey::Pubkey,
+        rent::Rent,
         signature::Keypair,
         system_instruction::transfer,
     },
@@ -37,9 +39,12 @@ use {
         get_associated_token_address_with_program_id,
         instruction::create_associated_token_account_idempotent,
     },
-    spl_token::instruction::{
-        close_account,
-        sync_native,
+    spl_token::{
+        instruction::{
+            close_account,
+            sync_native,
+        },
+        state::Account as TokenAccount,
     },
     std::str::FromStr,
 };
@@ -467,6 +472,28 @@ impl Svm {
                 e
             ))
         })
+    }
+
+    pub fn get_user_amount_to_wrap(
+        amount_user: u64,
+        user_mint_user_balance: u64,
+        token_account_initialization_configs: &TokenAccountInitializationConfigs,
+    ) -> u64 {
+        let number_of_atas_paid_by_user = [
+            &token_account_initialization_configs.user_ata_mint_user,
+            &token_account_initialization_configs.user_ata_mint_searcher,
+        ]
+        .iter()
+        .filter(|&&config| matches!(config, TokenAccountInitializationConfig::UserPayer))
+        .count();
+
+        std::cmp::min(
+            amount_user,
+            user_mint_user_balance.saturating_sub(
+                number_of_atas_paid_by_user as u64
+                    * Rent::default().minimum_balance(TokenAccount::LEN),
+            ),
+        )
     }
 
     /// Adjusts the bid amount in the case where the amount that needs to be provided by the searcher is specified and the fees are in the user token.